requirements_ubuntu18.txt
Flask==2.0.3
Flask-WTF==1.0.1
email-validator==1.3.1
python-dotenv==0.20.0
Werkzeug==2.0.3
WTForms==2.3.3
Bootstrap-Flask==2.0.2
Flask-Mail==0.9.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
click==8.0.4
dnspython==2.2.1
idna==3.4
